Split tagController into tagController and tagService and add appropriate unit tests
